Move selector to selectors/profile
