+++ src/selectors/url-state.js
@@ -257,33 +257,6 @@ export const getFormattedMetaInfoString: Selector<string | null> = (state) => {
   return formatMetaInfoString(profile.meta);
 };
 
-/** Does the profile have implementation data? */
-export const getProfileUsesFrameImplementation: Selector<boolean> = (state) => {
-  const { profile } = state.profileView;
-  if (!profile) {
-    return true;
-  }
-  return profile.meta.doesNotUseFrameImplementation !== true;
-};
-
-/** Should the "Look up the function name on Searchfox" menu entry be hidden? */
-export const getShouldDisplaySearchfox: Selector<boolean> = (state) => {
-  const { profile } = state.profileView;
-  if (!profile) {
-    return true;
-  }
-  return profile.meta.sourceCodeIsNotOnSearchfox !== true;
-};
-
-/* Hide the stack type of frames in context menus? */
-export const getProfileUsesMultipleStackTypes: Selector<boolean> = (state) => {
-  const { profile } = state.profileView;
-  if (!profile) {
-    return true;
-  }
-  return profile.meta.usesOnlyOneStackType !== true;
-};
-
 /**
  * Get just the file name from the zip file path, if it exists.
  */
